pyproject.toml
[project]
name = "psycho-agent"
version = "0.1.0"
description = "LangGraph-based semantic POMDP psychotherapy agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.26",
    "langchain-core>=0.3.12",
    "openai>=1.51.0",
    "tenacity>=9.0.0",
    "pydantic>=2.9.2",
    "neo4j>=5.25.0",
    "sentence-transformers>=3.1.1",
    "faiss-cpu>=1.8.0.post1",
    "orjson>=3.10.7",
    "numpy>=2.1.2",
    "memgpt>=0.7.3",
    "typer>=0.12.5",
    "rich>=13.9.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.9"
]

[build-system]
requires = ["setuptools>=75.2.0"]
build-backend = "setuptools.build_meta"
